adds bun stack

+++ templates/prettier/.template.sh
@@ -3,4 +3,8 @@
 npm pkg set scripts.format="prettier --write ."
 npm pkg set scripts.lint="prettier --check ."
 
-npm install --save-dev prettier
+if [ -f "bun.lock" ]; then
+  bun install --dev prettier
+else
+  npm install --save-dev prettier
+fi
